Add or modify specs for researches files
