+++ packages/yoastseo/spec/languageProcessing/researches/imageCountInTextSpec.js
@@ -1,5 +1,5 @@
-import imageCountFunction from "../../src/languages/legacy/researches/imageCountInText.js";
-import Paper from "../../src/values/Paper";
+import imageCountFunction from "../../../src/languageProcessing/researches/imageCount.js";
+import Paper from "../../../src/values/Paper";
 
 describe( "Counts images in an text", function() {
 	let imageCount;
